requirements.txt
altgraph==0.17.4
et_xmlfile==2.0.0
importlib_metadata==8.7.0
macholib==1.16.3
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
pycryptodome==3.23.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
scapy==2.6.1
tkinterdnd2==0.4.3
zipp==3.23.0
